+++ catalogue_graph/src/ingestor/models/display/short_description.py
