+++ catalogue_graph/src/ingestor/transformers/work_display_transformer.py
@@ -27,26 +27,23 @@ from ingestor.models.display.production_event import DisplayProductionEvent
 from ingestor.models.display.relation import (
     DisplayRelation,
 )
+from ingestor.models.display.short_description import derive_short_description
 from models.pipeline.archive_category import ArchiveCategory
 from models.pipeline.concept import Concept
 from models.pipeline.identifier import Identified
 from utils.sort import natural_sort_key
 
-from .short_description import derive_short_description
 from .work_base_transformer import WorkBaseTransformer
 
 
 class DisplayWorkTransformer(WorkBaseTransformer):
     @property
     def short_description(self) -> str | None:
-        # """Collection roots, excluding TEI manuscripts."""
-        # TEI_IDENTIFIER_TYPE = "tei-manuscript-id"
-        # if not self.is_collection_root:
-        #     return None
-        # if self.state.source_identifier.identifier_type.id == TEI_IDENTIFIER_TYPE:
-        #     return None
-        """Archive collection roots only, which is the population the browse cards need."""
-        if not self.is_collection_root or self.archive is None:
+        """Collection roots, excluding TEI manuscripts."""
+        TEI_IDENTIFIER_TYPE = "tei-manuscript-id"
+        if not self.is_collection_root:
+            return None
+        if self.state.source_identifier.identifier_type.id == TEI_IDENTIFIER_TYPE:
             return None
         return derive_short_description(self.data.description)
